Some preprocessing and modeling code
Separated the preprocessing (copy) part of the code and the parameter/modeling part of the code. Also, in the modeling part, there are AIC and BIC comparisons now. The nice little graph codes (radiating error) are also added at the end of modeling code. Just for error checking. In conclusion, more streamlined.